Resctrl collector should be aware of existing mon groups.

+++ resctrl/collector.go
@@ -20,6 +20,7 @@ package resctrl
 import (
 	"fmt"
 	"os"
+	"path/filepath"
 	"sync"
 	"time"
 
@@ -155,9 +156,15 @@ func (c *collector) Destroy() {
 func (c *collector) clear() error {
 	// Not allowed to remove root or undefined resctrl directory.
 	if c.id != rootContainer && c.resctrlPath != "" {
-		err := os.RemoveAll(c.resctrlPath)
-		if err != nil {
-			return fmt.Errorf("couldn't clear mon_group: %v", err)
+		// Remove only own prepared mon group.
+		base := filepath.Base(c.resctrlPath)
+		if len(base) > len(monGroupPrefix) {
+			if base[:len(monGroupPrefix)] == monGroupPrefix {
+				err := os.RemoveAll(c.resctrlPath)
+				if err != nil {
+					return fmt.Errorf("couldn't clear mon_group: %v", err)
+				}
+			}
 		}
 	}
 	return nil
